master data modules creation

+++ master-data/src/main/java/com/nextuple/pe/masterdata/domain/entity/ItemEntity.java
@@ -89,7 +89,7 @@ public class ItemEntity {
   private String weightUom;
 
   @Column(name = "processing_time")
-  private Long processingTime;
+  private Double processingTime;
 
   @Column(name = "cost")
   private String cost;
