+++ master-data/src/main/java/com/nextuple/pe/masterdata/domain/outbound/ItemResponse.java
@@ -34,7 +34,7 @@ public class ItemResponse implements Serializable {
   private String volumeUom;
   private Double weight;
   private String weightUom;
-  private Long processingTime;
+  private Double processingTime;
   private String cost;
   private Boolean isHazmat;
   private Long leadTime;
